fix:edit profile window
